mindquest-pm-agent/requirements.txt
anthropic>=0.25.0
asyncio
aiofiles>=23.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.1.0
schedule>=1.2.0
gitpython>=3.1.0
pygithub>=2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jinja2>=3.1.0
pyyaml>=6.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0